msm8996: QCamera2: Extend vendor specific format to gralloc
Issue:
In the current implementation the stream
format used by camera HAL for vendor specific camera formats is not passed to gralloc and it is pre decided.
This can lead to format mismatch if display and camera code is not from the same branch.

Fix:
Defined macros
and exposed them to gralloc to query vendor specific camera formats.

Change-Id: I4a5b632a7d988ec50641007ebcd1e3cdcd1278ae
CRs-fixed: 1014678